First steps at turning internals of SpineML simulator into library and calling this from very minimal command line application
