+++ include/spineml/simulator/input.h
@@ -35,7 +35,7 @@ namespace Input
 class Base
 {
 public:
-    virtual ~Base(){}
+    virtual ~Base();
 
     //----------------------------------------------------------------------------
     // Declared virtuals
@@ -80,7 +80,7 @@ protected:
 
     //----------------------------------------------------------------------------
     // Protected API
-    //----------------------------------------------------------------------------
+    //----------------------------------------------------------------------------InputValue
     void injectSpike(unsigned int neuronID);
 
     void uploadSpikes();
